+++ src/strip/strip.cpp
@@ -11,7 +11,7 @@
 #include <invader/tag/parser/parser.hpp>
 #include <invader/file/file.hpp>
 
-int strip_tag(const std::filesystem::path &file_path, bool preprocess) {
+int strip_tag(const std::filesystem::path &file_path) {
     // Open the tag
     auto tag = Invader::File::open_file(file_path);
     if(!tag.has_value()) {
@@ -24,7 +24,7 @@ int strip_tag(const std::filesystem::path &file_path, bool preprocess) {
     try {
         const auto *header = reinterpret_cast<const Invader::HEK::TagFileHeader *>(tag->data());
         Invader::HEK::TagFileHeader::validate_header(header, tag->size());
-        file_data = Invader::Parser::ParserStruct::parse_hek_tag_file(tag->data(), tag->size(), preprocess)->generate_hek_tag_data(header->tag_fourcc, true);
+        file_data = Invader::Parser::ParserStruct::parse_hek_tag_file(tag->data(), tag->size())->generate_hek_tag_data(header->tag_fourcc, true);
     }
     catch(std::exception &e) {
         eprintf_error("Error: Failed to strip %s: %s", file_path.string().c_str(), e.what());
@@ -48,7 +48,6 @@ int main(int argc, char * const *argv) {
     options.emplace_back("info", 'i', 0, "Show license and credits.");
     options.emplace_back("tags", 't', 1, "Use the specified tags directory.", "<dir>");
     options.emplace_back("fs-path", 'P', 0, "Use a filesystem path for the tag path if specifying a tag.");
-    options.emplace_back("preprocessor", 'p', 0, "Save the result of the tag preprocessor rather than just a strip. This is to allow easier tag comparison.");
     options.emplace_back("all", 'a', 0, "Strip all tags in the tags directory.");
 
     static constexpr char DESCRIPTION[] = "Strips extra hidden data from tags.";
@@ -58,7 +57,6 @@ int main(int argc, char * const *argv) {
         std::optional<std::filesystem::path> tags;
         bool use_filesystem_path = false;
         bool all = false;
-        bool preprocess = false;
     } strip_options;
 
     auto remaining_arguments = Invader::CommandLineOption::parse_arguments<StripOptions &>(argc, argv, options, USAGE, DESCRIPTION, 0, 1, strip_options, [](char opt, const std::vector<const char *> &arguments, auto &strip_options) {
@@ -79,16 +77,11 @@ int main(int argc, char * const *argv) {
             case 'a':
                 strip_options.all = true;
                 break;
-            case 'p':
-                strip_options.preprocess = true;
-                break;
         }
     });
     if(!strip_options.tags.has_value()) {
         strip_options.tags = "tags";
     }
-
-    bool preprocess = strip_options.preprocess;
 
     if(remaining_arguments.size() == 0) {
         if(!strip_options.all) {
@@ -99,14 +92,14 @@ int main(int argc, char * const *argv) {
         std::size_t success = 0;
         std::size_t total = 0;
 
-        auto recursively_strip_dir = [&total, &success, &preprocess](const std::filesystem::path &dir, auto &recursively_strip_dir) -> void {
+        auto recursively_strip_dir = [&total, &success](const std::filesystem::path &dir, auto &recursively_strip_dir) -> void {
             for(auto i : std::filesystem::directory_iterator(dir)) {
                 if(i.is_directory()) {
                     recursively_strip_dir(i, recursively_strip_dir);
                 }
                 else if(i.is_regular_file()) {
                     total++;
-                    success += strip_tag(i.path().string().c_str(), preprocess) == EXIT_SUCCESS;
+                    success += strip_tag(i.path().string().c_str()) == EXIT_SUCCESS;
                 }
             }
         };
@@ -130,6 +123,6 @@ int main(int argc, char * const *argv) {
             file_path = std::filesystem::path(*strip_options.tags) / Invader::File::halo_path_to_preferred_path(remaining_arguments[0]);
         }
         std::string file_path_str = file_path.string();
-        return strip_tag(file_path_str.c_str(), preprocess);
+        return strip_tag(file_path_str.c_str());
     }
 }